Fix import for newer DragonRuby versions
